Conflicts:
	FractalLandscapeGenerator/src/flg/rendering/RenderLandscape.java

+++ FractalLandscapeGenerator/src/flg/algorithm/test/DiamondSquareTest.java
@@ -2,16 +2,17 @@ package flg.algorithm.test;
 
 import flg.algorithm.DiamondSquare;
 
+
 public class DiamondSquareTest {
 	public static void runTest() {
-		double[][] points = new double[5][5];
+		float[][] points = new float[5][5];
 		
-		points[0][0] = 13d;
-		points[0][4] = -12.232d;
-		points[4][0] = 0.1213d;
-		points[4][4] = -120.d;
+		points[0][0] = 13f;
+		points[0][4] = -12.232f;
+		points[4][0] = 0.1213f;
+		points[4][4] = -120.0f;
 		
-		DiamondSquare.applyDiamondSquare(points, 0.8d);
+		DiamondSquare.applyDiamondSquare(points, 0.8f);
 		
 		for (int i = 0; i < 5; ++i) {
 			for (int j = 0; j < 5; ++j) {
